+++ tests/unit/test_models/test_submodels/test_electrolyte/test_stefan_maxwell/test_conductivity/test_surface_form/test_full_surface_form_stefan_maxwell_conductivity.py
@@ -27,6 +27,8 @@ class TestFull(unittest.TestCase):
             "Negative electrode interfacial current density": a_n,
             "Electrolyte potential": pybamm.Concatenation(a_n, a_s, a_p),
             "Negative electrode temperature": a_n,
+            "Separator temperature": a_s,
+            "Positive electrode temperature": a_p,
         }
         icd = " interfacial current density"
         reactions = {